+++ src/index.d.ts
@@ -1,0 +1,9 @@
+import Manager from "./Manager";
+
+declare const manager: Manager;
+
+export { Manager, manager };
+
+// Enums
+import { RepositoryType, NodeType, BrowserType, SyntaxType } from "./Manager";
+export { RepositoryType, NodeType, BrowserType, SyntaxType };